Make separated dlma handler
- #33
- Done.
