+++ backend/src/common.h
@@ -14,5 +14,12 @@ typedef struct _app {
   json_t* j_conf;
 } app;
 
+typedef enum _E_USE
+{
+  E_USE_NO = 0,
+  E_USE_OK = 1,
+} E_USE;
+
+
 
 #endif /* BACKEND_SRC_COMMON_H_ */
