+++ backend/src/ob_dl_handler.h
@@ -24,26 +24,9 @@ typedef enum _E_DL_STATUS_T
 //  E_DL_QUEUEING   = 2,  ///< not in used.
 } E_DL_STATUS_T;
 
-typedef enum _E_DL_USE
-{
-  E_DL_USE_NO = 0,
-  E_DL_USE_OK = 1,
-} E_DL_USE;
-
-
-
 int get_current_dialing_dl_cnt(const char* camp_uuid, const char* dl_table);
 int get_dial_num_point(json_t* j_dl_list, json_t* j_plan);
 int get_dial_try_cnt(json_t* j_dl_list, int dial_num_point);
-
-bool validate_ob_dlma(json_t* j_data);
-json_t* create_ob_dlma(json_t* j_dlma);
-json_t* update_ob_dlma(const json_t* j_dlma);
-json_t* delete_ob_dlma(const char* uuid);
-json_t* get_ob_dlmas_all(void);
-json_t* get_ob_dlmas_all_uuid(void);
-json_t* get_ob_dlma(const char* uuid);
-json_t* get_deleted_ob_dlma(const char* uuid);
 
 json_t* get_ob_dl(const char* uuid);
 json_t* get_ob_dls_uuid_by_dlma_count(const char* dlma_uuid, int count);
