+++ backend/src/ob_http_handler.c
@@ -26,6 +26,7 @@
 #include "ob_dialing_handler.h"
 #include "ob_dl_handler.h"
 #include "ob_plan_handler.h"
+#include "ob_dlma_handler.h"
 
 extern evhtp_t* g_htp;
